projects/stock-trading/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
PyJWT>=2.0.0
requests>=2.31.0
httpx>=0.24.0
python-multipart>=0.0.6